modified the Drawer w/functional component
